requirements.txt
# Web Framework
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-Cors==4.0.0
gunicorn==21.2.0
Werkzeug==2.3.7

# AI Integration
openai==1.3.7
langchain==0.0.346
langchain-openai==0.0.2.post1
tiktoken==0.5.1

# Blockchain
web3==6.0.0
vyper==0.3.9
eth-account==0.8.0
eth-typing==3.2.0

# Data Processing
requests==2.31.0

# Utility
python-dotenv==1.0.0
pydantic==2.1.1
